Add of the the cfn-dia library to projen
Ok so the process to add a library and make it a part of projen CI is
handle all through the `.projen` file. If you it to be a call out in the
CI and it is a script you can add it as a script in `.projen` OR if you
want the script with a description then add it as a Task.

Tasks can be powerful! There are even subtasks! [Refer here for documentation](https://github.com/projen/projen/blob/main/docs/tasks.md)